test: add comprehensive calculator unit tests
- Add unit tests for EMI, SIP, GST, and RD calculators
- Update build configuration and refactor UI currency handling

+++ app/src/main/java/com/loanmaster/pro/domain/model/CalculatorModels.kt
@@ -58,7 +58,7 @@ data class MonthlyAmortization(
 )
 
 data class FdYearBreakdown(
-    val year: Int,
+    val year: Double,
     val openingBalance: Double,
     val interestEarned: Double,
     val closingBalance: Double
